Abstracts Program struct to enable common use
