+++ .integration/tests/dpc_testnet1.rs
@@ -26,6 +26,7 @@ use snarkvm_utilities::{to_bytes_le, FromBytes, ToBytes};
 use rand::SeedableRng;
 use rand_chacha::ChaChaRng;
 use std::{
+    ops::Deref,
     sync::Arc,
     time::{SystemTime, UNIX_EPOCH},
 };
@@ -89,7 +90,7 @@ fn dpc_testnet1_integration_test() {
     for j in 0..Testnet1Parameters::NUM_OUTPUT_RECORDS {
         output_records.push(
             Record::new_output(
-                &dpc.noop_program,
+                dpc.noop_program.deref(),
                 recipient.address,
                 false,
                 10,
@@ -209,7 +210,7 @@ fn test_testnet1_transaction_authorization_serialization() {
     for j in 0..Testnet1Parameters::NUM_OUTPUT_RECORDS {
         output_records.push(
             Record::new_output(
-                &dpc.noop_program,
+                dpc.noop_program.deref(),
                 test_account.address,
                 false,
                 10,
@@ -284,7 +285,7 @@ fn test_testnet1_dpc_execute_constraints() {
     for j in 0..Testnet1Parameters::NUM_OUTPUT_RECORDS {
         output_records.push(
             Record::new_output(
-                &dpc.noop_program,
+                dpc.noop_program.deref(),
                 new_account.address,
                 false,
                 10,
